+++ WebApi/ModelsAndDtoes/Project_Incompleted/CreateProject_IncompletedApiDto.cs
@@ -17,7 +17,7 @@ namespace WebApi.ModelsAndDtoes.Project_Incompleted
         //نرخ بازگشت سرمایه طرح
         public int ReturnInvestmentRate { get; set; }
         //میزان سرمایه مورد تقاضا
-        public int AmountCapitalDemand { get; set; }
+        public long AmountCapitalDemand { get; set; }
         //آدرس
         [Required]
         public CreateAddressApiDto Address { get; set; }
@@ -48,15 +48,15 @@ namespace WebApi.ModelsAndDtoes.Project_Incompleted
         public bool JustificationPlan { get; set; }
         //واگذاری
         //قیمت پیشنهادی
-        public int ProposedPrice { get; set; }
+        public long ProposedPrice { get; set; }
         //مشارکت
         //کل سرمایه مورد نیاز طرح
-        public int RequiredCapitalPlan { get; set; }
+        public long RequiredCapitalPlan { get; set; }
         //مشارکت
         //در چه سرفصل‌هایی نیاز به مشارکت دارید؟
         public string WhatTopicsNeedParticipate { get; set; }
         //خالص ارزش فعلی پروژه
-        public int NetPresentValue { get; set; }
+        public long NetPresentValue { get; set; }
         //درصد پیشرفت کلی طرح تا کنون
         [Range(0, 100)]
         public int OverallProgressPercent { get; set; }
