+++ WebApi/ModelsAndDtoes/Project_UnderCapacity/CreateProject_UnderCapacityApiDto.cs
@@ -17,7 +17,7 @@ namespace WebApi.ModelsAndDtoes.Project_UnderCapacity
         //نرخ بازگشت سرمایه طرح
         public int ReturnInvestmentRate { get; set; }
         //میزان سرمایه مورد تقاضا
-        public int AmountCapitalDemand { get; set; }
+        public long AmountCapitalDemand { get; set; }
         //آدرس
         [Required]
         public CreateAddressApiDto Address { get; set; }
@@ -46,13 +46,13 @@ namespace WebApi.ModelsAndDtoes.Project_UnderCapacity
         public bool HaveFinancialStatement { get; set; }
         //واگذاری
         //قیمت پیشنهادی
-        public int ProposedPrice { get; set; }
+        public long ProposedPrice { get; set; }
         //مشارکت
         //میزان مشارکت مورد نیاز
         public int AmountOfParticipationRequired { get; set; }
         //مشارکت
         //کل سرمایه مورد نیاز طرح
-        public int RequiredCapitalPlan { get; set; }
+        public long RequiredCapitalPlan { get; set; }
         //مشارکت
         //در چه سرفصل‌هایی نیاز به مشارکت دارید؟
         public string WhatTopicsNeedParticipate { get; set; }
