+++ WebApi/ModelsAndDtoes/Project__ReadyToUse/CreateProject_ReadyToUseApiDto.cs
@@ -18,7 +18,7 @@ namespace WebApi.ModelsAndDtoes.Project__ReadyToUse
         //نرخ بازگشت سرمایه طرح
         public int ReturnInvestmentRate { get; set; }
         //میزان سرمایه مورد تقاضا
-        public int AmountCapitalDemand { get; set; }
+        public long AmountCapitalDemand { get; set; }
         //آدرس
         public CreateAddressDto Address { get; set; }
         /// <summary>
@@ -46,15 +46,15 @@ namespace WebApi.ModelsAndDtoes.Project__ReadyToUse
         public bool JustificationPlan { get; set; }
         //واگذاری
         //قیمت پیشنهادی
-        public int ProposedPrice { get; set; }
+        public long ProposedPrice { get; set; }
         //مشارکت
         //کل سرمایه مورد نیاز طرح
-        public int RequiredCapitalPlan { get; set; }
+        public long RequiredCapitalPlan { get; set; }
         //مشارکت
         //در چه سرفصل‌هایی نیاز به مشارکت دارید؟
         public string WhatTopicsNeedParticipate { get; set; }
         //خالص ارزش فعلی پروژه
-        public int NetPresentValue { get; set; }
+        public long NetPresentValue { get; set; }
         //درصد پیشرفت کلی طرح تا کنون
         [Range(0, 100)]
         public int OverallProgressPercent { get; set; }
